+++ Algo-Project/Item.h
@@ -2,8 +2,8 @@
 
 struct Item {
 	int data; //vertex
-	int key; //d[vertex]
+	float key; //d[vertex]
 
 	Item() : data(-1), key(-1) {}
-	Item(int d, int k) : data(d), key(k) {}
+	Item(int d, float k) : data(d), key(k) {}
 };
